iconv() returns an unsigned value so direct comparison with -1 is invalid. (dm)
git-svn-id: svn://mielke.cc/main/brltty@5583 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/charset.c
@@ -59,7 +59,7 @@ ret convert##name (from f) { \
     to t; \
     to *tp = &t; \
     size_t ts = sizeof(t); \
-    if (iconv(iconv##name, (void *)&fp, &fs, (void *)&tp, &ts) != -1) return t; \
+    if (iconv(iconv##name, (void *)&fp, &fs, (void *)&tp, &ts) != (size_t)-1) return t; \
     logMessage(LOG_DEBUG, "iconv (" #from " -> " #to ") error: %s", strerror(errno)); \
   } \
   return eof; \
